+++ NEWS.md
@@ -6,7 +6,8 @@
 - fixing python dependency (aiohappyeyeballs)
 - adding parameter `plot_n_word_random` and `legend_number_colour` in textPlot. 
 - removed `nltk` warning when running the functions requiring pyhon.
-- anchouring group word embeddings in the `textProjection()` function. 
+- anchouring group word embeddings in the `textProjection()` function.
+- adding cohen's d to the output of the `textProjection()` function
 
 # text 1.3.4
 - harmonizing wordclouds with topics-package